+++ usr.sbin/nsd/simdzone/CHANGELOG.md
@@ -5,6 +5,17 @@ All notable changes to simdzone will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.2.4] - 2025-12-04
+
+### Added
+
+### Fixed
+
+- Correct lengths for GOST R 34.10-2012 and SM3 delegation signer (DS) digest
+  algorithms
+- Require the AMTRELAY relay field to be `.` for the no gateway relay type as
+  specified by RFC 8777 (#257)
+
 ## [0.2.3] - 2025-09-03
 
 ### Added